refactor: springboot3 升级改造
1. 修复不兼容的 API
2. 依赖升级
3. servlet 依赖替换
4. mybatis-plus API 修改(plus javadoc 感觉就是一坨屎, 标注了过时 API, 也不说使用什么类替换, 兼容性升级这一块做的太差了, 所以后期应该不会考虑升级 mybatis-plus 了)

+++ blen-kernel-common/src/main/java/dev/dong4j/zeka/kernel/common/base/Bridge.java
@@ -3,12 +3,11 @@ package dev.dong4j.zeka.kernel.common.base;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import dev.dong4j.zeka.kernel.common.api.BaseCodes;
 import dev.dong4j.zeka.kernel.common.assertion.IAssert;
-import org.jetbrains.annotations.NotNull;
-
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * <p>Description: 由于返回类型不同和范型问题, 需要使用此类做桥接. 使用抽象类，避免后期添加顶层接口影响业务代码 </p>
@@ -280,7 +279,7 @@ public abstract class Bridge<D extends BaseDTO<? extends Serializable>> {
      * @return int int
      * @since 2.1.0
      */
-    public int counts(@NotNull IAssert assertions) {
+    public long counts(@NotNull IAssert assertions) {
         return assertions.wrapper(() -> service.counts());
     }
